change type of name  to float

+++ db/migrate/20210615161038_create_comments.rb
@@ -4,7 +4,7 @@ class CreateComments < ActiveRecord::Migration[5.2]
       t.string :content
       t.references :user, foreign_key: true
       t.references :post, foreign_key: true
-      t.integer :rate, null: false, default: 0
+      t.float :rate, null: false, default: 0
 
       t.timestamps
     end
